+++ app/core/routers/Products.py
@@ -1,5 +1,4 @@
 from fastapi import APIRouter, Body, HTTPException, status, Depends
-from fastapi.encoders import jsonable_encoder
 from typing import List
 
 from ..models.Product import ProductBase, ProductUpdate
@@ -7,20 +6,6 @@ from ..services import Products
 from ..services import Users
 
 router = APIRouter()
-
-
-@router.post("/", response_description="Add a new product")
-async def add_product(
-        current_user: dict = Depends(Users.get_current_user),
-        product_data: ProductBase = Body()):
-    if not current_user["isAdmin"]:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="You are not authorized to access this endpoint. Please contact your administrator."
-        )
-
-    product = jsonable_encoder(product_data)
-    return await Products.add_products(product_data=product)
 
 
 @router.get("/", response_description="Get all products that are in stock", response_model=List[ProductBase])
@@ -50,42 +35,4 @@ async def get_products_by_keyword(keyword: str):
     return products
 
 
-@router.patch("/{_id}/shelf-or-resell", response_description="Shelf or resell product")
-async def shelf_or_resell_product(
-        _id: str,
-        current_user: dict = Depends(Users.get_current_user)):
-    if not current_user["isAdmin"]:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="You are not authorized to access this endpoint. Please contact your administrator."
-        )
-
-    is_product_shelved_or_resold = await Products.shelf_or_resell_product(_id=_id)
-    if not is_product_shelved_or_resold:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Product is not in our database or not in stock")
-
-    return is_product_shelved_or_resold
-
-
-@router.patch("/{_id}", response_description="Update product", response_model=ProductBase)
-async def update_product(
-        _id: str,
-        product_data: ProductUpdate = Body(),
-        current_user: dict = Depends(Users.get_current_user)
-    ):
-    if not current_user["isAdmin"]:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="You are not authorized on this endpoint. Please contact your administrator."
-        )
-
-    is_item_updated = await Products.update_product(_id=_id, product_data=product_data)
-    if not is_item_updated:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Product is not in our database or not in stock")
-
-    return is_item_updated
