+++ app/core/services/Products.py
@@ -6,13 +6,6 @@ from ..models.Product import ProductBase, ProductUpdate
 products_collection = database.products_collection
 
 
-async def add_products(product_data: dict):
-    new_product = products_collection.insert_one(product_data)
-    created_product = products_collection.find_one({"_id": new_product.inserted_id})
-
-    return created_product
-
-
 async def get_all_products_available() -> List[ProductBase]:
     return list(products_collection.find({"isAvailable": True}))
 
@@ -27,33 +20,5 @@ async def get_available_product_by_id(_id: str) -> ProductBase:
 
 async def get_products_by_keyword(keyword: str) -> List[ProductBase]:
     return list(products_collection.find({"$text": {"$search": keyword}, "isAvailable": True}))
-
-
-async def update_product(_id: str, product_data: ProductUpdate):
-    product = await get_product_by_id(_id=_id)
-    if product is None:
-        return False
-
-    product["name"] = product_data.name
-    product["description"] = product_data.description
-    product["categories"] = product_data.categories
-    product["sku"] = product_data.sku
-    product["quantity"] = product_data.quantity
-    product["price"] = product_data.price
-    product["dateUpdated"] = datetime.now()
-
-    products_collection.update_one({"_id": _id}, {"$set": product})
-    return await get_product_by_id(_id=_id)
-
-
-async def shelf_or_resell_product(_id: str):
-    product = await get_product_by_id(_id=_id)
-    if product is None:
-        return False
-
-    product["isAvailable"] = not product["isAvailable"]
-    product["dateUpdated"] = datetime.now()
-    products_collection.update_one({"_id": _id}, {"$set": product})
-    return await get_product_by_id(_id=_id)
